+++ internal/kubernetes/job.go
@@ -1,17 +1,12 @@
 package kubernetes
 
 import (
-	"bytes"
 	"context"
-	"fmt"
-	"os"
 
 	batchv1 "k8s.io/api/batch/v1"
 	apiv1 "k8s.io/api/core/v1"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/cli-runtime/pkg/genericclioptions"
-	cp "k8s.io/kubectl/pkg/cmd/cp"
 )
 
 // RunJob will run job with specified parameters
@@ -104,19 +99,4 @@ func (k8s *Client) DeleteJob(name string) error {
 				PropagationPolicy: &propagationPolicy,
 			},
 		)
-}
-
-// Copy file from pod to local file
-func (k8s *Client) Copy(podName, podFilePath, localFilePath string) error {
-	ioStreams := genericclioptions.IOStreams{
-		In:     &bytes.Buffer{},
-		Out:    &bytes.Buffer{},
-		ErrOut: os.Stdout,
-	}
-	opts := cp.NewCopyOptions(ioStreams)
-	opts.Clientset = k8s.Clientset
-	opts.ClientConfig = k8s.Config
-	from := fmt.Sprintf("%s/%s:%s", k8s.Namespace, podName, podFilePath)
-
-	return opts.Run([]string{from, localFilePath})
 }
